refactor: rename the folder and update import
